--- grafana/doc.go
@@ -1,2 +1,0 @@
-// Package grafana contains a client to work with Grafana GRPC server API.
-package grafana
