+++ grpcapi/doc.go
@@ -1,0 +1,2 @@
+// Package grpcapi contains a client to work with Grafana plugin GRPC server API.
+package grpcapi
