+++ grpcapi/dto.go
@@ -1,4 +1,4 @@
-package grafana
+package grpcapi
 
 // PublishResult returned from Live server. This is empty at the moment,
 // but can be extended with fields later.
